update bean
Change LongId type to Integer

+++ src/main/java/com/fjcorona/springboot/user/endpoint/springbootuserendpoint/models/User.java
@@ -9,7 +9,7 @@ public class User {
 
 	@Id
 	@Column(name = "id")
-	private Long id;
+	private Integer id;
 	
 	@Column(name = "name")
 	private String name;
@@ -27,7 +27,7 @@ public class User {
 
 	}
 
-	public User(Long id, String name, String lastname, Integer age, Integer active) {
+	public User(Integer id, String name, String lastname, Integer age, Integer active) {
 		super();
 		this.id = id;
 		this.name = name;
@@ -36,11 +36,11 @@ public class User {
 		this.active = active;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
